Changes in scoring program. Previously, the function 'get_metric' didn't work for metrics defined in libscores because the function was in libscores itself.
