+++ starting_kit/scoring_output/scores.html
@@ -1,3 +1,3 @@
-<pre>======= Set 1 (Iris_test): accuracy_score(set1_score)=0.371428571429 =======
-<pre>======= Set 2 (Iris_train): accuracy_score(set2_score)=0.457142857143 =======
-<pre>======= Set 3 (Iris_valid): accuracy_score(set3_score)=0.314285714286 =======
+<pre>======= Set 1 (Iris_test): bac_multiclass(set1_score)=0.000000000000 =======
+<pre>======= Set 2 (Iris_train): bac_multiclass(set2_score)=0.000000000000 =======
+<pre>======= Set 3 (Iris_valid): bac_multiclass(set3_score)=0.000000000000 =======
