+++ starting_kit/scoring_output/scores.txt
@@ -1,4 +1,4 @@
-set1_score: 0.371428571429
-set2_score: 0.457142857143
-set3_score: 0.314285714286
+set1_score: 0.000000000000
+set2_score: 0.000000000000
+set3_score: 0.000000000000
 Duration: 0
