--- starting_kit/utilities/iris_bundle_19-05-13-16-15/competition.yaml
@@ -1,75 +1,0 @@
-# One-size-fit-all YAML file to build an M2AIC 2018-2019 challenge bundle
-
-# DO NOT MODIFY
-
-# Isabelle Guyon -- May 13, 2019
-
-title: Sample Competition  
-description: Sample competition M2AIC 2018-2019
-image: logo.jpg
-has_registration: False                 # We do not check participants, registration is free
-force_submission_to_leaderboard: True   # If false, participants must submit manually, do not change
-disallow_leaderboard_modifying: True    # If true, submissions cannot be changed
-enable_detailed_results: True           # True if html file written
-enable_forum: True                      # Forum active
-admin_names: guyon,tthomas63,eric,pavao # Admin people that can help you will have access to your competition
-competition_docker_image: codalab/codalab-legacy:py3
-end_date: null                          # Never ends
-
-html:
-  data: data.html
-  evaluation: evaluation.html
-  overview: overview.html
-  terms: rules.html
-  #notebook: README.html
-
-phases:
-  1:
-    phasenumber: 1
-    label: Development Phase
-    description: 'Development phase: tune your models and submit prediction results, trained model, or untrained model.'
-    start_date: 2018-11-15
-    is_scoring_only: False    
-    execution_time_limit: 500
-    max_submissions_per_day: 5   
-    force_best_submission_to_leaderboard: True      # Participants will see their best submission on the leaderboard
-    starting_kit: starting_kit.zip                  # The starting kit you prepared
-    ingestion_program: ingestion_program.zip        # The ingestion program (the same for both phases)
-    public_data: input_data.zip                     # Same as input data (available for download by the participants)
-    input_data: input_data.zip                      # The data used by the ingestion program (and the code of the participants) in both phases
-    scoring_program: scoring_program.zip            # The scoring program (the same for both phases)
-    reference_data: reference_data_1.zip            # The truth values (solution) for phase 1 used by the scoring program
-    color: green   
-  2:
-    phasenumber: 2
-    label: Final Phase
-    description: 'Final phase (no submission, your last submission from the previous phase is automatically forwarded).'
-    start_date: 2050-04-30
-    is_scoring_only: False 
-    execution_time_limit: 300
-    max_submissions_per_day: 5  
-    auto_migration: True    
-    ingestion_program: ingestion_program.zip        # The ingestion program (the same for both phases)
-    input_data: input_data.zip                      # The data used by the ingestion program (and the code of the participants) in both phases
-    scoring_program: scoring_program.zip            # The scoring program (the same for both phases)
-    reference_data: reference_data_2.zip            # The truth values (solution) for phase 1 used by the scoring program
-    color: purple
-
-leaderboard:
-    leaderboards:
-        Results: &RESULTS
-            label: RESULTS
-            rank: 1
-    columns:
-        set1_score:
-            leaderboard: *RESULTS
-            label: Prediction score
-            numeric_format: 4
-            rank: 1
-        Duration:
-            leaderboard: *RESULTS
-            label: Duration 
-            numeric_format: 2
-            rank: 2
-            
-
