--- starting_kit/utilities/iris_bundle_19-05-13-16-15/data.html
@@ -1,10 +1,0 @@
-
-<h1>Data</h1>
-
-<p>
-The data come from the Iris dataset from UCI repository. See the full description <a href="http://archive.ics.uci.edu/ml/datasets/Iris">here</a>.
-</p>
-<p> Go to the <b>FILES</b> tab to download the data and the starting kit. The starting kit contains a very small subset of the data
-for debug purposes. To prepare a challenge submission, you need to use the large dataset that you download separately.
-For the the large dataset, you do NOT have the labels!
-</p>
