--- starting_kit/utilities/iris_bundle_19-05-13-16-15/evaluation.html
@@ -1,31 +1,0 @@
-
-<h1>Evaluation</h1>
-
-<p>
-The problem is a multiclass classification problem. Each sample (an Iris) is characterized by its sepal and petal
-width and length (4 features). You must predict the Iris categories: setosa, virginica, or versicolor.
-<br>
-You are given for training a data matrix X_train of dimension num_training_samples x num_features 
-and an array y_train of labels of dimension num_training_samples. You must train a model
-which predicts the labels for two test matrices X_valid and X_test. 
-
-<br>
-There are 2 phases:
-<ul>
-<li> <b>Phase 1: development phase. </b> We provide you with labeled training data and unlabeled validation and test data.
-Make predictions for both datasets. However, you will receive feed-back on your performance on the validation set only. 
-The performance of your LAST submission will be displayed on the leaderboard.
-<li> <b>Phase 2: final phase.</b> You do not need to do anything. Your last submission of phase 1 will be automatically forwarded.
-Your performance on the test set will appear on the leaderboard when the organizers finish checking the submissions. 
-</ul>
-
-This sample competition allows you to submit either:
-<ul>
-<li> Only prediction results (no code).
-<li> A pre-trained prediction model. 
-<li> A prediction model that must be trained and tested. 
-</ul>
-
-The submissions are evaluated using the <b>accuracy</b> metric.
-</p>
-
