--- starting_kit/utilities/iris_bundle_19-05-13-16-15/overview.html
@@ -1,10 +1,0 @@
-<h1>Solve Fisher's Famous Iris Problem</h1>
-<h3>Brought to you by the ChaLearn</h3>
-<p>This is the well known Iris dataset from Fisher's classic paper (Fisher, 1936). The data set contains 3 classes of 50 instances each, where each class refers to a type of iris plant. 
-One class is linearly separable from the other 2; the latter are NOT linearly separable from each other. </p>
-
-<p><b>References and credits:</b>
-<br>R. A. Fisher. The use of multiple measurements in taxonomic problems. Annual Eugenics, 7, Part II, 179-188 (1936).
-<br>The competition protocol was designed by Isabelle Guyon.
-<br>The starting kit was adapted from an Jupyper notebook designed by Balazs Kegl for the <a href="http://www.ramp.studio/">RAMP platform</a>.
-<br>This challenge was generated using Chalab, a competition wizard designed by Laurent Senta.
