--- starting_kit/utilities/iris_bundle_19-05-13-16-15/rules.html
@@ -1,9 +1,0 @@
-
-<h1>Rules</h1>
-
-<p>
-Submissions must be made before the end of phase 1.
-You may submit 5 submissions every day and 100 in total.
-</p>
-
-This challenge is governed by the <a href="http://www.causality.inf.ethz.ch/GeneralChalearnContestRuleTerms.html">general ChaLearn contest rules</a>.
